v1.1.0: dockers ordered by alphabetical order
- Added changelog file
- separated menus in their own classes and files
