+++ CHANGELOG.md
@@ -1,0 +1,21 @@
+## [Unreleased]
+### Added
+### Changed
+### Removed
+### Fixes
+
+## [1.1.0] - 2022-05-22
+### Added
+- Changelog file
+### Changed
+- Separated both menus in their own files dMenu and rmenu
+### Fixes
+- Fixed dockers not being sorted by alphabetical order in the menu
+
+## [1.0.0] - 2022-05-18
+### Added
+- Creation of 2 menus, resources and dockers in the main menu
+- List of dockers outside of the dockers submenu krita provides
+- License in files
+### Changed
+- Changed name from resources-and-dockers-menu to RD Menus
